NAV-27173: Rydder opp i høyremeny pakker
