+++ src/frontend/sider/Fagsak/Behandling/Høyremeny/Brev/Brev.tsx
@@ -5,8 +5,8 @@ import { useNavigate } from 'react-router';
 import { Button, Modal, VStack } from '@navikt/ds-react';
 
 import { Brevskjema } from './Brevskjema';
-import { useBrukerContext } from '../../../sider/Fagsak/BrukerContext';
-import { useFagsakContext } from '../../../sider/Fagsak/FagsakContext';
+import { useBrukerContext } from '../../../BrukerContext';
+import { useFagsakContext } from '../../../FagsakContext';
 
 interface Props {
     onIModalClick: () => void;
