+++ src/frontend/sider/Fagsak/Behandling/Høyremeny/Brev/useBrevModul.ts
@@ -5,16 +5,16 @@ import { feil, ok, useFelt, useSkjema, Valideringsstatus } from '@navikt/familie
 
 import type { ISelectOptionMedBrevtekst } from './typer';
 import { Brevmal } from './typer';
-import { useBehandlingContext } from '../../../sider/Fagsak/Behandling/context/BehandlingContext';
-import { Behandlingstype, BehandlingÅrsak, type IBehandling } from '../../../typer/behandling';
-import { BehandlingKategori } from '../../../typer/behandlingstema';
-import type { IManueltBrevRequestPåBehandling } from '../../../typer/dokument';
-import type { IGrunnlagPerson } from '../../../typer/person';
-import { PersonType } from '../../../typer/person';
-import type { IBarnMedOpplysninger } from '../../../typer/søknad';
-import { Målform } from '../../../typer/søknad';
-import type { IFritekstFelt } from '../../../utils/fritekstfelter';
-import { genererIdBasertPåAndreFritekster, lagInitiellFritekst } from '../../../utils/fritekstfelter';
+import { Behandlingstype, BehandlingÅrsak, type IBehandling } from '../../../../../typer/behandling';
+import { BehandlingKategori } from '../../../../../typer/behandlingstema';
+import type { IManueltBrevRequestPåBehandling } from '../../../../../typer/dokument';
+import type { IGrunnlagPerson } from '../../../../../typer/person';
+import { PersonType } from '../../../../../typer/person';
+import type { IBarnMedOpplysninger } from '../../../../../typer/søknad';
+import { Målform } from '../../../../../typer/søknad';
+import type { IFritekstFelt } from '../../../../../utils/fritekstfelter';
+import { genererIdBasertPåAndreFritekster, lagInitiellFritekst } from '../../../../../utils/fritekstfelter';
+import { useBehandlingContext } from '../../context/BehandlingContext';
 
 export const hentMuligeBrevmalerImplementering = (behandling: IBehandling): Brevmal[] => {
     const brevmaler: Brevmal[] = Object.keys(Brevmal) as Brevmal[];
