+++ src/frontend/sider/Fagsak/Behandling/Høyremeny/Ikoner/Ikon.tsx
