+++ src/frontend/typer/dokument.ts
@@ -1,7 +1,7 @@
 import type { BehandlingKategori } from './behandlingstema';
 import type { Målform } from './søknad';
-import type { Brevmal, Informasjonsbrev } from '../komponenter/Hendelsesoversikt/BrevModul/typer';
 import type { SkjemaBrevmottaker } from '../komponenter/Saklinje/Meny/LeggTilEllerFjernBrevmottakere/useBrevmottakerSkjema';
+import type { Brevmal, Informasjonsbrev } from '../sider/Fagsak/Behandling/Høyremeny/Brev/typer';
 
 export interface IManueltBrevRequestPåBehandling {
     mottakerIdent: string;
